Initial Game State Machine Check in
GameStateMachine now holds a list of game states that the user can add
in the editor (in the Game Scene).

The GameStateMachine will parse these state game objects and add them to
a state dictionary.

Game States can send a messege to the state machine to change the state
when their state is finished