ChampionDBManager, Party/PartyMember/Being and Sprite Data
My champion DB manager is grabbing literally everything from the API
everytime you run the app which is probably not the greatest way of
doing it. I'll have to look into somehow grabbing the data at run time
tomorrow. I noticed that my API calls on my dashboard were really high.
In order to have the champion data, you'll need to have the
ChampionDBManager on a gameobject in your scene.